+++ src/test/java/org/scify/jedai/blockbuilding/TestAllMethodsDirtyER.java
@@ -52,14 +52,14 @@ public class TestAllMethodsDirtyER {
             System.out.println("Existing Duplicates\t:\t" + duplicatePropagation.getDuplicates().size());
 
             for (BlockBuildingMethod blbuMethod : BlockBuildingMethod.values()) {
-                float time1 = System.currentTimeMillis();
+                long time1 = System.currentTimeMillis();
 
                 System.out.println("\n\nCurrent blocking metohd\t:\t" + blbuMethod);
                 IBlockBuilding blockBuildingMethod = BlockBuildingMethod.getDefaultConfiguration(blbuMethod);
 
                 System.out.println("Block Building...");
                 List<AbstractBlock> blocks = blockBuildingMethod.getBlocks(profiles, null);
-                float time2 = System.currentTimeMillis();
+                long time2 = System.currentTimeMillis();
 
                 BlocksPerformance blStats = new BlocksPerformance(blocks, duplicatePropagation);
                 blStats.setStatistics();
