+++ src/test/java/org/scify/jedai/configuration/HolisticRandomConfigurationDER.java
@@ -82,7 +82,7 @@ public class HolisticRandomConfigurationDER {
             int bestIteration = 0;
             float bestFMeasure = 0;
             for (int j = 0; j < NO_OF_TRIALS; j++) {
-                float time1 = System.currentTimeMillis();
+                long time1 = System.currentTimeMillis();
 
                 bb.setNextRandomConfiguration();
                 final List<AbstractBlock> blocks = bb.getBlocks(profiles);
@@ -111,7 +111,7 @@ public class HolisticRandomConfigurationDER {
                 ec.setNextRandomConfiguration();
                 final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-                float time2 = System.currentTimeMillis();
+                long time2 = System.currentTimeMillis();
 
                 final StringBuilder matchingWorkflowConf = new StringBuilder();
                 matchingWorkflowConf.append(bb.getMethodConfiguration());
@@ -135,7 +135,7 @@ public class HolisticRandomConfigurationDER {
             System.out.println("\nBest Iteration\t:\t" + bestIteration);
             System.out.println("Best FMeasure\t:\t" + bestFMeasure);
 
-            float time1 = System.currentTimeMillis();
+            long time1 = System.currentTimeMillis();
 
             bb.setNumberedRandomConfiguration(bestIteration);
             final List<AbstractBlock> blocks = bb.getBlocks(profiles);
@@ -155,7 +155,7 @@ public class HolisticRandomConfigurationDER {
             ec.setNumberedRandomConfiguration(bestIteration);
             final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-            float time2 = System.currentTimeMillis();
+            long time2 = System.currentTimeMillis();
 
             final StringBuilder matchingWorkflowConf = new StringBuilder();
             matchingWorkflowConf.append(bb.getMethodConfiguration());
