+++ src/test/java/org/scify/jedai/configuration/StepByStepGridConfigurationDER.java
@@ -47,8 +47,8 @@ import java.util.List;
  */
 public class StepByStepGridConfigurationDER {
 
-    static int getTotalComparisons(List<AbstractBlock> blocks) {
-        int originalComparisons = 0;
+    static long getTotalComparisons(List<AbstractBlock> blocks) {
+        long originalComparisons = 0;
         for (AbstractBlock block : blocks) {
             originalComparisons += block.getNoOfComparisons();
         }
@@ -91,7 +91,7 @@ public class StepByStepGridConfigurationDER {
             // local optimization of Block Building 
             float bestA = 0;
             int bestIteration = 0;
-            int originalComparisons = profiles.size() * (profiles.size() - 1) / 2;
+            long originalComparisons = profiles.size() * (profiles.size() - 1) / 2;
             for (int j = 0; j < bb.getNumberOfGridConfigurations(); j++) {
                 bb.setNumberedGridConfiguration(j);
                 final List<AbstractBlock> originalBlocks = bb.getBlocks(profiles);
@@ -236,7 +236,7 @@ public class StepByStepGridConfigurationDER {
             System.out.println("\nBest Outer Iteration\t:\t" + bestOuterIteration);
             System.out.println("Best FMeasure\t:\t" + bestFMeasure);
 
-            float time1 = System.currentTimeMillis();
+            long time1 = System.currentTimeMillis();
 
             em.setNumberedGridConfiguration(bestOuterIteration);
             final SimilarityPairs sims = em.executeComparisons(finalBlocks);
@@ -244,7 +244,7 @@ public class StepByStepGridConfigurationDER {
             ec.setNumberedGridConfiguration(bestInnerIteration);
             final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-            float time2 = System.currentTimeMillis();
+            long time2 = System.currentTimeMillis();
 
             final StringBuilder matchingWorkflowConf = new StringBuilder();
             matchingWorkflowConf.append(bb.getMethodConfiguration());
