+++ src/test/java/org/scify/jedai/configuration/version2_1/HolisticRandomConfigurationDER.java
@@ -95,7 +95,7 @@ public class HolisticRandomConfigurationDER {
         matchingWorkflowConf.append("\n").append(em.getMethodConfiguration());
         matchingWorkflowConf.append("\n").append(ec.getMethodConfiguration());
 
-        float time1 = System.currentTimeMillis();
+        long time1 = System.currentTimeMillis();
 
         final List<AbstractBlock> blocks = bb.getBlocks(profiles);
         final List<AbstractBlock> purgedBlocks = bp1.refineBlocks(blocks);
@@ -104,7 +104,7 @@ public class HolisticRandomConfigurationDER {
         final SimilarityPairs sims = em.executeComparisons(finalBlocks);
         final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-        float time2 = System.currentTimeMillis();
+        long time2 = System.currentTimeMillis();
 
         final ClustersPerformance clp = new ClustersPerformance(clusters, duplicatePropagation);
         clp.setStatistics();
